completed week-8
Very hard do implement in my opinion, I won't lie I had some assistance using AI to help me with .split .replace. and .trim. It was very very new to me so I believe having assistance was okay for this part. It took me a very long time because at first I thought it was my other codes that were the problem as to why the meals were not showing only to find out it was because of this. I had fun with this challenge though!